+++ EmojiThrower/GameOverScene.swift
@@ -11,20 +11,20 @@ class GameOverScene: SKScene {
         
         super.init(size: size)
 
-        backgroundColor = SKColor.white()
+        backgroundColor = .white
 
         let message = won ? "You Won!" : "You Lose!"
 
         let label = SKLabelNode(fontNamed: "Helvetica")
         label.text = message
         label.fontSize = 40
-        label.fontColor = SKColor.black()
+        label.fontColor = .black
         label.position = CGPoint(x: size.width/2, y: size.height/2)
         addChild(label)
 
         run(SKAction.sequence([
             SKAction.wait(forDuration: 3.0),
-            SKAction.run() {
+            SKAction.run {
                 let reveal = SKTransition.flipHorizontal(withDuration: 0.2)
                 self.restartGame(size: size, transition: reveal)
             }
@@ -43,7 +43,7 @@ class GameOverScene: SKScene {
         return true
     }
     
-    required init(coder aDecoder: NSCoder) {
+    required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
 }